Updated library path with install_name_tool
This allows -lplumed to work correctly even when DYLIB_LIBRARY_PATH
is not set.

Looking at the other dylibs installed by MacPorts
I realized all of them include the install path in this way,
so I think this is standard.

We might consider including this on "make install" for mac
also when not using MacPorts.  As a quick fix, I add it here.